+++ release/ewoh-0.6.0-rc2/ewoh-spark-app/server/modules/system/system.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { SystemConfigController } from './system.controller';
+import { FeatureFlagsController, SystemConfigController } from './system.controller';
 import { SystemService } from './system.service';
 
 @Module({
-  controllers: [SystemConfigController],
+  controllers: [SystemConfigController, FeatureFlagsController],
   providers: [SystemService],
   exports: [SystemService],
 })